backend/requirements.txt
fastapi
starlette>=0.41.3
uvicorn[standard]
pydantic
python-dotenv

boto3

yt_dlp
moviepy
TikTokApi
scenedetect
opencv-python

google-genai
pinecone[grpc]
openai

re
